Refactor: 온보딩 타입 파일 네이밍 변경
