+++ src/entities/user-onboarding.entity.ts
@@ -1,6 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsNumber, IsNotEmpty, IsString } from 'class-validator';
-import { userOnboardingStatusType } from 'src/utils/types';
+import { userOnboardingStatusType } from 'src/utils/onBoardingtypes';
 import { Column, Entity, JoinColumn, OneToOne } from 'typeorm';
 import { BaseEntity } from './base.entity';
 import { User } from './user.entity';
