+++ src/utils/onBoardingtypes.ts
@@ -1,4 +1,5 @@
 export type userOnboardingStatusType = 'WAIT' | 'DONE_WAIT' | 'DONE';
+
 export type userOnboardingType =
   | 'makeFirstNameCard'
   | 'shareMyNameCard'
